feat(grammar): extend typeref to variable and attribute types, addressing known ambiguity
extend the typeref allowing variables and attributes to be derived
types.

Introduces precedence into the grammar for the known ambiguity
x : attr = ...
In this instance x should be treated as a type for global attribute
attr, and NOT as a variable name x with variable attribute attr.
Correct interpretation relies on knowing what derived types have been
declared.