feat(grammar): allow attributes to appear in types and dimensions
Allows attributes to appear in types and note sections alongside derived
types and variable declarations as per the spec.

Does not allow attributes to appear in data, even though the cdl spec
implies this, as ncgen raises an error. Will open this with NetCDF User
Group.

Based on the syntax a generic attribute can appear in any section, but
in reality only global attributes can appear in types or dimensions.
Consider tightening this at a later point.